Setting a simpler task path structure, Adding some layout to the task types, and fixing a ui bug in the rating task
