+++ website/src/middleware.ts
@@ -4,5 +4,5 @@ export { default } from "next-auth/middleware";
  * Guards all pages under `/grading` and redirects them to the sign in page.
  */
 export const config = {
-  matcher: ["/grading/:path*"],
+  matcher: ["/create/:path*", "/evaluate/:path*"],
 };
