+++ website/src/pages/index.tsx
@@ -1,18 +1,15 @@
-import { useSession } from "next-auth/react";
-
+import { Button, Input, Stack } from "@chakra-ui/react";
 import Head from "next/head";
 import Link from "next/link";
-import { Button, Input, Stack } from "@chakra-ui/react";
+import { useSession } from "next-auth/react";
 
-import { CallToAction } from "../components/CallToAction";
-import { Faq } from "../components/Faq";
-import { Footer } from "../components/Footer";
-import { Header } from "../components/Header";
-import { Hero } from "../components/Hero";
+import { CallToAction } from "src/components/CallToAction";
+import { Faq } from "src/components/Faq";
+import { Footer } from "src/components/Footer";
+import { Header } from "src/components/Header";
+import { Hero } from "src/components/Hero";
 
-import styles from "../styles/Home.module.css";
-
-export default function Home() {
+const Home = () => {
   const { data: session } = useSession();
 
   if (!session) {
@@ -48,13 +45,15 @@ export default function Home() {
       <Header />
       <main className="h-3/4  z-0 bg-white flex flex-col items-center justify-center gap-2">
         <Button size="lg" colorScheme="blue" className="drop-shadow">
-          <Link href="/grading/grade-output">Rate a prompt and output now</Link>
+          <Link href="/evaluate/rate_summary">Rate a Summary</Link>
         </Button>
         <Button size="lg" colorScheme="blue" className="drop-shadow">
-          <Link href="/summarize/story">Summarize a story</Link>
+          <Link href="/create/summarize_story">Summarize a story</Link>
         </Button>
       </main>
       <Footer />
     </>
   );
-}
+};
+
+export default Home;
